ADD: SOFA-MISC_STATIC_LIBRARY option in CMake to build static Sofa libraries.
Former-commit-id: c43dd4f59e9b78ad61eadc525cce4448d941df8c

+++ framework/sofa/helper/pooledEigen/Pool.cpp
@@ -11,7 +11,7 @@ namespace internal {
 
 boost::pool<>* Pool::get_pool(size_t N)
 {
-	const size_t n = (size_t) ceilf(logf(N) * (1.0/M_LN2));
+	const size_t n = (size_t) ceilf(logf((float)N) * (float)(1.0/M_LN2));
 
 	if( gs_pools[n] == NULL )
 		gs_pools[n] = new boost::pool<>(1 << n);
